\name{ExampleData}
\alias{ExampleData}
\docType{data}
\title{
ExampleData
}
\description{
Artificial dataset to be used with \code{samp}.
}
\usage{data("ExampleData")}
\format{
  A data frame with 40000 observations on the following 6 variables.
  \describe{
    \item{\code{stratum}}{a character vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{npop}}{a numeric vector}
    \item{\code{nsamp}}{a numeric vector}
    \item{\code{prn}}{a numeric vector}
    \item{\code{size}}{a numeric vector}
  }
}
\examples{
data(ExampleData)
}
\keyword{datasets}
