\name{ExampleData}
\alias{ExampleData}
\docType{data}
\title{
ExampleData
}
\description{
Artificial business dataset to be used with \code{samp}.
}
\usage{data("ExampleData")}
\format{
  A data frame with 40030 observations on the following 8 variables.
  \describe{
    \item{\code{nace}}{a factor with levels \code{4611} \code{4612} \code{4613} \code{4614} \code{4615} \code{4616} \code{4617} \code{4618} \code{4619} \code{4621} \code{4622} \code{4623} \code{4624} \code{4631} \code{4632} \code{4633} \code{4634} \code{4635} \code{4636} \code{4637} \code{4638} \code{4639} \code{4641} \code{4642} \code{4643} \code{4644} \code{4645} \code{4646} \code{4647} \code{4648} \code{4649} \code{4651} \code{4652} \code{4661} \code{4662} \code{4663} \code{4664} \code{4665} \code{4666} \code{4669} \code{4671} \code{4672} \code{4673} \code{4674} \code{4675} \code{4676} \code{4677} \code{4690}}
    \item{\code{id}}{a numeric vector}
    \item{\code{N_emp}}{a numeric vector}
    \item{\code{inv}}{a numeric vector}
    \item{\code{turn}}{a numeric vector}
    \item{\code{nsamp}}{an array}
    \item{\code{ntot}}{a numeric vector}
    \item{\code{prn}}{a numeric vector}
  }
}

\examples{
data(ExampleData)
}
\keyword{datasets}
