\name{ExampleData}
\alias{ExampleData}
\docType{data}
\title{
ExampleData
}
\description{
Artificial dataset to be used with \code{samp} and \code{transformprn}.
}
\usage{data("ExampleData")}
\format{
  A data frame with 40000 observations on the following 6 variables.
  \describe{
    \item{\code{stratum}}{a character vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{npopul}}{a numeric vector}
    \item{\code{nsample}}{a numeric vector}
    \item{\code{rands}}{a numeric vector}
    \item{\code{sizeM}}{a numeric vector}
  }
}
\examples{
data(ExampleData)
}
\keyword{datasets}
