% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{ExampleData}
\format{
## `ExampleData`

A data frame with 40,000 rows and 6 columns:
\describe{
 \item{\code{stratum}}{a character vector}
 \item{\code{id}}{a numeric vector}
 \item{\code{npopul}}{a numeric vector}
 \item{\code{nsample}}{a numeric vector}
 \item{\code{rands}}{a numeric vector}
 \item{\code{sizeM}}{a numeric vector}
}
}
\source{
Ad-hoc simulation in base R.
}
\usage{
ExampleData
}
\description{
Artificial dataset to be used with \code{samp} and \code{transformprn}.
}
\seealso{
\link{prnsamplr}, \link{samp}, \link{srs}, \link{pps},
\link{transformprn}
}
\keyword{datasets}
